Refactor in progress in structure and content
