+++ lib/constants/colors_constants.dart
@@ -1,9 +1,53 @@
 import 'package:flutter/material.dart';
 
 class ColorsConstants {
-  static const Color primary = Color(0xff094067);
-  static const Color secondary = Color(0xff90b4ce);
-  static const Color tertiary = Color(0xff3da9fc);
   static const Color neutral = Color(0xff5f6c7b);
-  static const Color error = Color(0xffef4565);
+
+  static const Color primaryLight = Color(0xff094067);
+  static const Color secondaryLight = Color(0xff00639d);
+  static const Color tertiaryLight = Color(0xff695779);
+  static const Color errorLight = Color(0xffba1a1a);
+  static const Color backgroundLight = Color(0xfffdfbff);
+  static const Color outlineLight = Color(0xff72777f);
+  static const Color onPrimaryLight = Color(0xffffffff);
+  static const Color onSecondaryLight = Color(0xffffffff);
+  static const Color onTertiaryLight = Color(0xffffffff);
+  static const Color onErrorLight = Color(0xffffffff);
+  static const Color onBackgroundLight = Color(0xff001b3d);
+  static const Color primaryContainerLight = Color(0xffcfe5ff);
+  static const Color secondaryContainerLight = Color(0xffcfe5ff);
+  static const Color tertiaryContainerLight = Color(0xfff0dbff);
+  static const Color errorContainerLight = Color(0xffffdad6);
+  static const Color surfaceLight = Color(0xfffdfbff);
+  static const Color surfaceVariantLight = Color(0xffdee3eb);
+  static const Color onPrimaryContainerLight = Color(0xff001d34);
+  static const Color onSecondaryContainerLight = Color(0xff001d33);
+  static const Color onTertiaryContainerLight = Color(0xff241532);
+  static const Color onErrorContainerLight = Color(0xff410002);
+  static const Color onSurfaceLight = Color(0xff001b3d);
+  static const Color onSurfaceVariantLight = Color(0xff42474e);
+
+  static const Color primaryDark = Color(0xff99cbff);
+  static const Color secondaryDark = Color(0xff98cbff);
+  static const Color tertiaryDark = Color(0xffd4bee6);
+  static const Color errorDark = Color(0xffffb4ab);
+  static const Color backgroundDark = Color(0xff001b3d);
+  static const Color outlineDark = Color(0xff8c9199);
+  static const Color onPrimaryDark = Color(0xff003355);
+  static const Color onSecondaryDark = Color(0xff003354);
+  static const Color onTertiaryDark = Color(0xff392a49);
+  static const Color onErrorDark = Color(0xff690005);
+  static const Color onBackgroundDark = Color(0xffd6e3ff);
+  static const Color primaryContainerDark = Color(0xff004a78);
+  static const Color secondaryContainerDark = Color(0xff004a77);
+  static const Color tertiaryContainerDark = Color(0xff514060);
+  static const Color errorContainerDark = Color(0xff93000a);
+  static const Color surfaceDark = Color(0xff001b3d);
+  static const Color surfaceVariantDark = Color(0xff42474e);
+  static const Color onPrimaryContainerDark = Color(0xffcfe5ff);
+  static const Color onSecondaryContainerDark = Color(0xffcfe5ff);
+  static const Color onTertiaryContainerDark = Color(0xfff0dbff);
+  static const Color onErrorContainerDark = Color(0xffffdad6);
+  static const Color onSurfaceDark = Color(0xffd6e3ff);
+  static const Color onSurfaceVariantDark = Color(0xffc2c7cf);
 }
