+++ lib/presentation/buttons/custom_outlined_button.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+
+class CustomOutlinedButton extends StatelessWidget {
+  final Function onPressed;
+  final String text;
+  final bool isFilled;
+
+  const CustomOutlinedButton({
+    Key? key,
+    required this.onPressed,
+    required this.text,
+    this.isFilled = false,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return OutlinedButton(
+        style: ButtonStyle(
+          shape: MaterialStateProperty.all(
+              RoundedRectangleBorder(borderRadius: BorderRadius.circular(10))),
+        ),
+        onPressed: () => onPressed(),
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
+          child: Text(
+            text,
+          ),
+        ));
+  }
+}
