+++ lib/presentation/cards/white_card.dart
