+++ lib/presentation/shared/widgets/Logo.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:store/constants/colors_constants.dart';
+
+class Logo extends StatelessWidget {
+  const Logo({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.only(top: 20),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          const Icon(Icons.photo,
+              color: ColorsConstants.tertiaryLight, size: 30),
+          const SizedBox(width: 10),
+          Text(
+            // size.width > 540 ? 'Tienda' : '', TODO: check the animation
+            '',
+            style: GoogleFonts.roboto(
+              color: Colors.white,
+              fontSize: 15,
+              fontWeight: FontWeight.bold,
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
