+++ lib/presentation/shared/widgets/notification_indicator.dart
