+++ lib/presentation/views/dashboard/dashboard_view.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:store/constants/colors_constants.dart';
+
+class DashboardView extends StatelessWidget {
+  const DashboardView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Icon(Icons.photo, color: ColorsConstants.tertiaryLight, size: 50),
+    );
+  }
+}
