+++ lib/router/handlers/auth_handlers.dart
@@ -1,9 +1,9 @@
 import 'package:fluro/fluro.dart';
 import 'package:provider/provider.dart';
 import 'package:store/provider/auth_provider.dart';
-import 'package:store/ui/views/auth/login_view.dart';
-import 'package:store/ui/views/auth/recovery_view.dart';
-import 'package:store/ui/views/dashboard/dashboard_view.dart';
+import 'package:store/presentation/views/auth/login_view.dart';
+import 'package:store/presentation/views/auth/recovery_view.dart';
+import 'package:store/presentation/views/dashboard/dashboard_view.dart';
 
 class AuthHandlers {
   static Handler login = Handler(handlerFunc: (context, params) {
