+++ lib/router/handlers/dashboard_handlers.dart
@@ -1,8 +1,8 @@
 import 'package:fluro/fluro.dart';
 import 'package:provider/provider.dart';
 import 'package:store/provider/auth_provider.dart';
-import 'package:store/ui/views/auth/login_view.dart';
-import 'package:store/ui/views/dashboard/dashboard_view.dart';
+import 'package:store/presentation/views/auth/login_view.dart';
+import 'package:store/presentation/views/dashboard/dashboard_view.dart';
 
 class DashboardHandler {
   static Handler dashboard = Handler(handlerFunc: (context, params) {
