+++ lib/router/handlers/no_page_found_handlers.dart
@@ -1,5 +1,5 @@
 import 'package:fluro/fluro.dart';
-import 'package:store/ui/views/no_page_found_view.dart';
+import 'package:store/presentation/views/no_page_found_view.dart';
 
 class NoPageFoundHandlers {
   static Handler noPageFound = Handler(handlerFunc: (context, params) {
