+++ lib/router/handlers/users_handlers.dart
@@ -2,10 +2,10 @@ import 'package:fluro/fluro.dart';
 import 'package:provider/provider.dart';
 import 'package:store/provider/auth_provider.dart';
 import 'package:store/provider/menu_provider.dart';
-import 'package:store/ui/views/auth/login_view.dart';
-import 'package:store/ui/views/dashboard/users/users_create_view.dart';
-import 'package:store/ui/views/dashboard/users/users_list_view.dart';
-import 'package:store/ui/views/dashboard/users/users_modify_view.dart';
+import 'package:store/presentation/views/auth/login_view.dart';
+import 'package:store/presentation/views/dashboard/users/users_create_view.dart';
+import 'package:store/presentation/views/dashboard/users/users_list_view.dart';
+import 'package:store/presentation/views/dashboard/users/users_modify_view.dart';
 
 class UsersHandlers {
   static Handler list = Handler(handlerFunc: (context, params) {
