--- lib/ui/shared/sidebar.dart
@@ -1,79 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:sidebarx/sidebarx.dart';
-import 'package:store/constants/colors_constants.dart';
-import 'package:store/provider/menu_provider.dart';
-
-class SideBar extends StatefulWidget {
-  const SideBar({super.key});
-
-  @override
-  State<SideBar> createState() => _SideBarState();
-}
-
-class _SideBarState extends State<SideBar> {
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final menuProvider = Provider.of<MenuProvider>(context);
-    return SidebarX(
-      controller: menuProvider.menuController,
-      theme: SidebarXTheme(
-        margin: const EdgeInsets.all(10),
-        decoration: BoxDecoration(
-          color: ColorsConstants.primary,
-          borderRadius: BorderRadius.circular(20),
-        ),
-        hoverColor: Colors.white.withOpacity(0.4),
-        textStyle: TextStyle(color: Colors.white.withOpacity(0.7)),
-        selectedTextStyle: const TextStyle(color: Colors.white),
-        itemTextPadding: const EdgeInsets.only(left: 10),
-        selectedItemTextPadding: const EdgeInsets.only(left: 10),
-        itemDecoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(10),
-          border: Border.all(color: ColorsConstants.primary.withOpacity(0.37)),
-        ),
-        selectedItemDecoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(10),
-          boxShadow: [
-            BoxShadow(
-              color: Colors.white.withOpacity(0.28),
-            )
-          ],
-        ),
-        iconTheme: IconThemeData(
-          color: Colors.white.withOpacity(0.7),
-          size: 20,
-        ),
-        selectedIconTheme: const IconThemeData(
-          color: Colors.white,
-          size: 20,
-        ),
-      ),
-      extendedTheme: const SidebarXTheme(
-        width: 200,
-        decoration: BoxDecoration(
-          color: ColorsConstants.primary,
-        ),
-      ),
-      footerDivider: Divider(
-        color: Colors.white.withOpacity(0.3),
-        height: 1,
-      ),
-      headerBuilder: (context, extended) {
-        return const SizedBox(
-          height: 100,
-          child: Padding(
-            padding: EdgeInsets.all(16.0),
-            child: Icon(Icons.image),
-          ),
-        );
-      },
-      items: menuProvider.getMenuItems,
-    );
-  }
-}
