+++ pubspec.yaml
@@ -25,6 +25,9 @@ dev_dependencies:
     sdk: flutter
   drift_dev: ^2.13.0
   build_runner: ^2.4.6
+  build_web_compilers: ^4.0.0
+  build: ^2.2.1
+
 
 flutter:
   uses-material-design: true
